move test, update version, and add type info
