+++ evergreen_lint/__init__.py
@@ -1,3 +1,3 @@
 """I exist to appease the linter."""
 
-__version__ = "0.1.2"
+__version__ = "0.1.4"  # Duplicated in pyproject.toml.
