+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "evergreen-lint"
-version = "0.1.4"
+version = "0.1.4"  # Duplicated in the top-level __init__.py.
 description = ""
 authors = ["Richard Samuels <richard.samuels@mongodb.com>"]
